[2ch.hk] Fix typo(#1062) + Fix preview for op-posts(#1066)

+++ src/Dollchan_Extension_Tools.es6.user.js
@@ -22,7 +22,7 @@
 'use strict';
 
 const version = '17.2.13.0';
-const commit = '4f6c5cd';
+const commit = '826766f';
 
 /*==[ DefaultCfg.js ]=========================================================================================
                                                 DEFAULT CONFIG
@@ -9773,16 +9773,17 @@ class Pview extends AbstractPost {
 		return Pview.top ? Pview.top.parent : null;
 	}
 	static show(parent, link) {
-		var rv, tNum = +(link.pathname.match(/.+?\/[^\d]*(\d+)/) || [,aib.getPostOfEl(link).tNum])[1],
-			pNum = +(link.textContent.trim().match(/\d+$/) || [tNum])[0],
-			isTop = !(parent instanceof Pview),
-			pv = isTop ? Pview.top : parent.kid;
+		const tNum = +(link.pathname.match(/.+?\/[^\d]*(\d+)/) || [,aib.getPostOfEl(link).tNum])[1];
+		const pNum = +(link.textContent.trim().match(/\d+$/) || [tNum])[0];
+		const isTop = !(parent instanceof Pview);
+		let pv = isTop ? Pview.top : parent.kid;
 		clearTimeout(Pview._delTO);
 		if(pv && pv.num === pNum) {
 			if(pv.kid) {
 				pv.kid.delete();
 			}
 			if(pv._link !== link) {
+				// If cursor hovers new link with the same number - move old preview here
 				pv._setPosition(link, Cfg.animation);
 				pv._link.classList.remove('de-link-parent');
 				link.classList.add('de-link-parent');
@@ -9796,6 +9797,7 @@ class Pview extends AbstractPost {
 			}
 			pv.parent = parent;
 		} else if(!Cfg.noNavigHidd || !pByNum.has(pNum) || !pByNum.get(pNum).hidden) {
+			// Show new preview under new link
 			if(pv) {
 				pv.delete();
 			}
@@ -9809,34 +9811,35 @@ class Pview extends AbstractPost {
 		return pv;
 	}
 	static updatePosition(scroll) {
-		var pv = Pview.top;
-		if(pv) {
-			var parent = pv.parent;
-			if(parent.omitted) {
+		let pv = Pview.top;
+		if(!pv) {
+			return;
+		}
+		const parent = pv.parent;
+		if(parent.omitted) {
+			pv.delete();
+			return;
+		}
+		if(parent.thr.loadCount === 1 && !parent.el.contains(pv._link)) {
+			const el = parent.ref.getElByNum(pv.num);
+			if(el) {
+				pv._link = el;
+			} else {
 				pv.delete();
 				return;
 			}
-			if(parent.thr.loadCount === 1 && !parent.el.contains(pv._link)) {
-				var el = parent.ref.getElByNum(pv.num);
-				if(el) {
-					pv._link = el;
-				} else {
-					pv.delete();
-					return;
-				}
+		}
+		const cr = parent.hidden ? parent : pv._link.getBoundingClientRect();
+		const diff = pv._isTop ? pv._offsetTop - window.pageYOffset - cr.bottom
+		                       : pv._offsetTop + pv.el.offsetHeight - window.pageYOffset - cr.top;
+		if(Math.abs(diff) > 1) {
+			if(scroll) {
+				scrollTo(window.pageXOffset, window.pageYOffset - diff);
 			}
-			var cr = parent.hidden ? parent : pv._link.getBoundingClientRect();
-			var diff = pv._isTop ? pv._offsetTop - (window.pageYOffset + cr.bottom)
-			                     : (pv._offsetTop + pv.el.offsetHeight) - (window.pageYOffset + cr.top);
-			if(Math.abs(diff) > 1) {
-				if(scroll) {
-					scrollTo(window.pageXOffset, window.pageYOffset - diff);
-				}
-				do {
-					pv._offsetTop -= diff;
-					pv.el.style.top = Math.max(pv._offsetTop, 0) + 'px';
-				} while((pv = pv.kid));
-			}
+			do {
+				pv._offsetTop -= diff;
+				pv.el.style.top = Math.max(pv._offsetTop, 0) + 'px';
+			} while((pv = pv.kid));
 		}
 	}
 	static _markLink(el, num) {
@@ -11405,7 +11408,7 @@ class MakabaPostsBuilder {
 		}
 
 		return `<div id="post-${ num }" class="post-wrapper">
-			<div class="post ${ i === '-1' ? 'oppost' : 'reply' }" id="post-body-${ num }" data-num="${ num }">
+			<div class="post ${ i === -1 ? 'oppost' : 'reply' }" id="post-body-${ num }" data-num="${ num }">
 				<div id="post-details-${ num }" class="post-details">
 					<input type="checkbox" name="delete" value="${ num }">
 					${ !data.subject ? '' :
@@ -13720,7 +13723,7 @@ function getImageBoard(checkDomains, checkEngines) {
 					Object.defineProperty(window, name, { value: emptyFn, writable: false, configurable: false });
 				}
 				fixGlobalFunc("$alert");
-				fixGlobalFunc("autorefresh_start"):
+				fixGlobalFunc("autorefresh_start");
 				fixGlobalFunc("linkremover");
 				fixGlobalFunc("scrollTo");
 				window.FormData = void 0;
